requirements.txt
pandas>=2.2.2
numpy>=1.26.4
SQLAlchemy>=2.0.30
PyMySQL>=1.1.1
python-dotenv>=1.0.1
pymongo>=4.7.2
tqdm>=4.66.5
folium>=0.16.0
Flask>=3.0.0

